Modify srand argument value.
# srand 인자값 수정.

+++ Mega_man_X4_WinAPI/Contents/ContentsCore.cpp
@@ -34,7 +34,7 @@ void ContentsCore::BeginPlay()
 	// ºÎ¸ð BeginPlay.
 	UEngineCore::BeginPlay();
 
-	srand(time(0));
+	srand(static_cast<unsigned int>(time(0)));
 
 	// ==========================================================================
